proyecto_final/dags/requirements.txt
apache-airflow==2.7.3
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
requests==2.31.0
mlflow==2.8.1
boto3==1.29.7
